fix opticalFlow.py

+++ opticalFlow.py
@@ -65,8 +65,8 @@ def main():
 
                 # draw the tracks
                 for i,(new,old) in enumerate(zip(good_new,good_old)):
-                    a,b = new.ravel()
-                    c,d = old.ravel()
+                    a,b = new.astype(int).ravel()
+                    c,d = old.astype(int).ravel()
                     mask = cv2.line(mask, (a,b),(c,d), color[i].tolist(), 2)
                     frame = cv2.circle(frame,(a,b),5,color[i].tolist(),-1)
                 img = cv2.add(frame,mask)
